+++ src/app/routes/loader/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Loader';
+export { default } from './RoutingLoader';
